Beta release with working update dialog on Windows

+++ installers/legacy-install.sh
@@ -18,6 +18,7 @@ Flags:
  --activate <project>            Activate a project when State Tool is correctly installed
  --activate-default <project>    Activate a project and make it the system default
  -h                              Show usage information (what you're currently reading)
+ -v <version-SHA>                The version of the State Tool to install
 EOF
 `
 
@@ -129,7 +130,7 @@ if [ -z "$TMPDIR" ]; then
 fi
 
 # Process command line arguments.
-while getopts "nb:t:e:c:f?h-:" opt; do
+while getopts "nb:t:e:c:v:f?h-:" opt; do
   case $opt in
   -)  # parse long options
     case ${OPTARG} in
@@ -162,6 +163,9 @@ while getopts "nb:t:e:c:f?h-:" opt; do
   n)
     NOPROMPT=true
     ;;
+  v)
+    VERSION=$OPTARG
+    ;;
   h|?)
     echo "${USAGE}"
     exit 0
@@ -178,6 +182,12 @@ fi
 
 if [ -n "$ACTIVATE" ] && [ -n "$ACTIVATE_DEFAULT" ]; then
   error "Flags --activate and --activate-default cannot be set at the same time."
+  exit 1
+fi
+
+if [ -z "$VERSION" ]; then
+  error "Please provide an argument for parameter '-v'. This installation script only installs specific State Tool versions."
+  info "Use 'https://platform.activestate.com/dl/cli/install.sh' to install the latest version of the State Tool."
   exit 1
 fi
 
@@ -248,20 +258,23 @@ if [ -f $TMPDIR/$TMPEXE ]; then
 fi
 
 fetchArtifact () {
-  info "Determining latest version..."
+  info "Fetching version info..."
   # Determine the latest version to fetch.
-  $FETCH $TMPDIR/$STATEJSON $STATEURL$STATEJSON || exit 1
-  VERSION=`cat $TMPDIR/$STATEJSON | grep -m 1 '"Version":' | awk '{print $2}' | tr -d '",'`
+  $FETCH $TMPDIR/$STATEJSON $STATEURL$VERSION/$STATEJSON
+  if [ $? -ne 0 ]; then
+    error "Failed to fetch info for version $VERSION.  Please check that the version string is valid."
+    exit 1
+  fi
   SUM=`cat $TMPDIR/$STATEJSON | grep -m 1 '"Sha256v2":' | awk '{print $2}' | tr -d '",'`
   rm $TMPDIR/$STATEJSON
 
-  if [ -z "$VERSION" ]; then
-    error "Unable to retrieve the latest version number"
+  info "Fetching version: $VERSION..."
+  # Fetch it.
+  $FETCH $TMPDIR/$STATEPKG ${STATEURL}${VERSION}/${STATEPKG}
+  if [ $? -ne 0 ]; then
+    error "Failed to download the State Tool archive.  Please try again later."
     exit 1
   fi
-  info "Fetching the latest version: $VERSION..."
-  # Fetch it.
-  $FETCH $TMPDIR/$STATEPKG ${STATEURL}${VERSION}/${STATEPKG} || exit 1
 
   # Extract the State binary after verifying its checksum.
   # Verify checksum.
